Switch to TypeScript (#30)
