+++ .editorconfig
@@ -1,0 +1,10 @@
+root = true
+
+[*.{ts,js,json}]
+charset = utf-8
+end_of_line = lf
+indent_size = 4
+indent_style = space
+insert_final_newline = false
+max_line_length = 120
+tab_width = 4
