+++ tsconfig.json
@@ -1,0 +1,15 @@
+{
+	"compilerOptions": {
+		"module": "commonjs",
+		"target": "ES2020",
+		"outDir": "out",
+		"lib": [
+			"ES2020",
+			"DOM",
+		],
+		"sourceMap": true,
+		"rootDir": "src",
+		"strict": true,
+		"resolveJsonModule": true
+	}
+}
